Verificacion/requirements.txt
appier==1.18.24
atomicwrites==1.3.0
attrs==19.1.0
beautifulsoup4==4.7.1
behave==1.2.6
behave-django==1.3.0
bs4==0.0.1
certifi==2019.3.9
chardet==3.0.4
coala-utils==0.5.1
colorama==0.4.1
DateTime==4.3
Django==2.2
future==0.17.1
idna==2.8
Jinja2==2.10.1
jmespath==0.9.4
lxml==4.3.3
MarkupSafe==1.1.1
mock==2.0.0
more-itertools==7.0.0
nltk==3.4.1
numpy==1.16.3
oauthlib==3.0.1
parse==1.12.0
parse-type==0.4.2
pbr==5.1.3
pluggy==0.9.0
py==1.8.0
PySocks==1.7.0
pytest==4.4.1
python-twitter==3.5
pytz==2019.1
PyYAML==5.1
render==1.0.0
requests==2.21.0
requests-oauthlib==1.2.0
selenium==3.141.0
six==1.12.0
soupsieve==1.9.1
sqlparse==0.3.0
stopwords==0.1.3
template==0.4.2
toml==0.10.0
tweepy==3.7.0
urllib3==1.24.2
zope.interface==4.6.0